Alarm Clock
This is an alarm clock coded for use with Arduino.  The setup includes three buttons: Minus, Plus, and Select.
These buttons are to be connected in between the 3.3v power source and the digital pin input, so when the button is pressed it allows voltage to flow into the Arduino module, sensing a 'HIGH' condition.  The alarm pin-out can be used with a small buzzer or with LEDs for a more colorful approach.